django/DRF_pjt/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
Django==4.2
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==34.0.0
fonttools==4.55.3
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.1
opendatasets==0.1.22
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
simplejwt==2.0.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
threadpoolctl==3.5.0
toposort==1.10
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
xgboost==2.1.3
